requirements.txt
# Core Flask Dependencies
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-dotenv>=1.0.0

# Database
PyMySQL>=1.1.0
cryptography>=41.0.0

# AI/ML
openai>=1.0.0
google-cloud-aiplatform>=1.35.0
google-auth>=2.23.0
vertexai>=0.0.1
langchain>=0.0.350
langchain-community>=0.0.10

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Scheduling
schedule>=1.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.2

# Markdown Processing
Markdown>=3.5.0

# Production Server
gunicorn>=21.2.0
eventlet>=0.33.3

# Monitoring & Logging
python-json-logger>=2.0.7

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Security
python-jose>=3.3.0
passlib>=1.7.4

# Rate Limiting
Flask-Limiter>=3.5.0

# CORS Support
Flask-Cors>=4.0.0

# Environment Management
python-decouple>=3.8